+++ src/WhereGroup/MetadorBundle/Component/MetadorController.php
@@ -150,7 +150,11 @@ class MetadorController extends Controller
         }
 
         // SAVE NEW ADDRESSES
-        $addresses = array_merge($p['responsiblePartyMetadata'], $p['responsibleParty']);
+        $addresses = array_merge(
+            $p['responsiblePartyMetadata'], 
+            $p['responsibleParty'],
+            $p['responsiblePartyDistributor']
+        );
 
         foreach($addresses as $row) {
             if(trim(@$row['organisationName']) == ""
@@ -187,6 +191,7 @@ class MetadorController extends Controller
                 $address->setVoice(@$row['voice']);
                 $address->setFacsimile(@$row['facsimile']);
                 $address->setOnlineResource(@$row['onlineResource']);
+                $address->setPositionName(@$row['positionName']);
 
                 $em->persist($address);
                 $em->flush();